requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
PyNaCl==1.5.0
